pyproject.toml
[project]
name = "to-radius-log"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "clickhouse-driver>=0.2.9",
    "pika>=1.3.2",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.1",
]
